requirements.txt
# Meow Blog Django 博客系统依赖包

# Django 核心框架
Django==5.2.6
python-dotenv==1.0.1

# Markdown 支持
markdown==3.5.1
Pygments==2.17.2  # 代码高亮支持

# 数据库支持
# SQLite（开发环境）- Django 内置支持
# MySQL（生产环境）
PyMySQL==1.1.1
cryptography==43.0.1
# mysqlclient==2.2.4
# pip install --only-binary=all mysqlclient
# PostgreSQL（生产环境推荐）
# psycopg2==2.9.9

# Web 服务器（生产环境）
gunicorn==22.0.0  # Unix/Linux 系统
waitress==2.1.2   # Windows 系统

# 日志和监控
django-request-logging==0.7.0  # 请求日志记录
django-extensions==3.2.3       # Django扩展工具

# 数据库管理工具
mysql-connector-python==8.2.0  # MySQL自动化配置脚本依赖

# 其他可选依赖
# Pillow==10.4.0  # 图片处理（如果需要头像上传功能）
# redis==5.1.1    # 缓存和会话存储
# celery==5.4.0   # 异步任务处理